Update HW2-CemalDagli.py
It is the second homework of the "Introduction of Python Programming Course" organized by GlobalAIHub. The main purpose of the script is to store the data of the job applicants in a dictionary. After storing the data of the job applicants in a dictionary, the data is converted into a list to ease the printing as well as for using another data type.